💄 `/delete-account`にデザインをあてる

+++ frontend/src/features/user/DeleteUser.tsx
@@ -11,6 +11,12 @@ const DeleteUser = () => {
       return
     }
 
+    const confirmation = window.confirm('アカウントを削除しますか？')
+
+    if (!confirmation) {
+      return
+    }
+
     try {
       const response = await axios.delete(`${railsApiUrl}/users/${session.user.email}`)
 
@@ -28,9 +34,12 @@ const DeleteUser = () => {
 
   if (session) {
     return (
-      <div className='hover:opacity-50 transition-all duration-100'>
-        <button onClick={() => handleDeleteUser()}>アカウントを削除する</button>
-      </div>
+      <button
+        onClick={() => handleDeleteUser()}
+        className='hover:opacity-50 transition-all duration-100 text-base bg-red-500 hover:bg-red-700 text-white font-notoSans py-2 px-4 rounded-full'
+      >
+        アカウントを削除する
+      </button>
     )
   }
   return null
